gcp-firestore-to-bq/requirements.txt
google-auth==1.34.0
google-cloud==0.34.0
google-cloud-bigquery==2.7.0
google-cloud-core==1.5.0
pandas==1.2.1
pyarrow==5.0.0
firebase-admin==6.0.1